Add base settings fragment
