+++ app/src/main/java/com/twilio/video/app/ui/settings/BandwidthProfileSettingsFragment.kt
@@ -1,34 +1,18 @@
 package com.twilio.video.app.ui.settings
 
-import android.content.Context
 import android.content.SharedPreferences
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.DialogFragment
 import androidx.preference.ListPreference
-import androidx.preference.Preference
-import androidx.preference.PreferenceFragmentCompat
 import com.twilio.video.app.R
 import com.twilio.video.app.data.NumberPreference
-import com.twilio.video.app.data.NumberPreferenceDialogFragmentCompat
 import com.twilio.video.app.data.Preferences
 import com.twilio.video.app.data.get
-import dagger.android.support.AndroidSupportInjection
 import javax.inject.Inject
 
-class BandwidthProfileSettingsFragment : PreferenceFragmentCompat() {
+class BandwidthProfileSettingsFragment : BaseSettingsFragment() {
 
     @Inject
     internal lateinit var sharedPreferences: SharedPreferences
-
-    private var previousTitle: CharSequence? = null
-
-    override fun onAttach(context: Context) {
-        AndroidSupportInjection.inject(this)
-        super.onAttach(context)
-    }
 
     override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
         // Add our preference from resources
@@ -71,28 +55,5 @@ class BandwidthProfileSettingsFragment : PreferenceFragmentCompat() {
                     Preferences.BANDWIDTH_PROFILE_TRACK_SWITCH_OFF_MODE_DEFAULT))
         (findPreference(Preferences.BANDWIDTH_PROFILE_TRACK_SWITCH_OFF_MODE) as ListPreference)
             .setValueIndex(bandwidthProfileTrackSwitchOffModeDefaultIndex)
-    }
-
-    override fun onDisplayPreferenceDialog(preference: Preference?) {
-        if (preference == null) {
-            return
-        }
-
-        // show custom dialog preference
-        if (preference is NumberPreference) {
-            val dialogFragment: DialogFragment?
-            dialogFragment = NumberPreferenceDialogFragmentCompat.newInstance(preference.key)
-
-            if (dialogFragment != null) {
-                dialogFragment.setTargetFragment(this, 0)
-                dialogFragment.show(requireFragmentManager(), BANDWIDTH_PROFILE_PREFERENCE_FRAGMENT_TAG)
-            }
-        } else {
-            super.onDisplayPreferenceDialog(preference)
-        }
-    }
-
-    companion object {
-        private const val BANDWIDTH_PROFILE_PREFERENCE_FRAGMENT_TAG = "BANDWIDTH_PROFILE_PREFERENCE_FRAGMENT_TAG"
     }
 }
